backend-cloud/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.5.0
flask-cors>=4.0.0
google-genai>=0.3.0